+++ src/machinable/project/project.py
@@ -9,7 +9,7 @@ import sys
 from ..config.loader import from_callable as load_config_from_callable
 from ..config.loader import from_file as load_config_file
 from ..config.loader import from_string as load_config_from_string
-from ..config.parser import parse_module_list
+from ..config.parser import parse_module_list, ModuleClass
 from ..core import Component as BaseComponent
 from ..core import Mixin as BaseMixin
 from ..core.component import FunctionalComponent