+++ tests/test_project/nodes/default_fallback.py
@@ -1,0 +1,6 @@
+from machinable import Component
+
+
+class DefaultFallback(Component):
+    def on_execute(self):
+        print("Default component", self.config)